+++ R/datasets.R
@@ -617,6 +617,25 @@
 #' \item{year}{Year of Sale}
 #' \item{region}{Region Where Sale Took Place}
 #' }
-"avocados" 
+"avocados"
+
+
+#' House Prices and Properties in Saratoga, New York
+#' 
+#' Random sample of 1057 houses taken from full Saratoga Housing Data (De Veaux)
+#' @format A data frame with 1057 observations on the following 8 variables
+#' \describe{
+#' \item{price}{price (US dollars)} 
+#' \item{living_area}{Living Area (square feet)}
+#' \item{bathrooms}{Number of Bathroom (half bathrooms have no shower or tub)}
+#' \item{bedrooms}{Number of Bedrooms}
+#' \item{fireplaces}{Number of Fireplaces}
+#' \item{lot_size}{Size of Lot (acres)}
+#' \item{age}{Age of House (years)}
+#' \item{fireplace}{Whether the house has a Fireplace}
+#' }
+#' @source Gathered from <https://docs.google.com/spreadsheets/d/1AY5eECqNIggKpYF3kYzJQBIuuOdkiclFhbjAmY3Yc8E/edit#gid=622599674>
+"saratoga_houses" 
+