+++ NEWS.md
@@ -8,7 +8,7 @@
 * Added `ma_traffic_2020_vs_2019` consisting of collisions information sourced from reports produced by the Massachusetts Traffic Data Management System.
 * Added `mario-kart-auction` dataset
 * Added `avocados` consisting of avocado prices dataset downloaded from the Hass Avocado Board website in May of 2018.
-
+* Added `saratoga_houses` random sample of 1057 houses taken from full Saratoga Housing Data. 
 
 
 # moderndive 0.5.3.9000